{{cookiecutter.project_slug}}/django/requirements/local.txt
{% raw %}
-r base.txt

django-debug-toolbar~=1.6.0
pdbpp
pylint
coverage
{% endraw %}
